+++ src/data_prep/__main__.py
@@ -1,0 +1,2 @@
+if __name__ == "__main__":
+    pass
