+++ src/data_prep/cli.py
@@ -1,0 +1,170 @@
+from data_prep.dialog_handler import DialogHandler
+from data_prep.precise_ops import (
+    BasicFileOperations,
+    TrainTestSplit,
+    PreciseModelingOperations,
+    GaussianNoiseHandler,
+)
+from get_base_model_flow import get_base_model_flow
+from data_prep.data_generation_flow import data_generation_flow
+from data_prep.further_data_generation_flow import further_data_generation_flow
+
+
+import sys
+import json
+
+# source_directory = str(input("Please enter the relative path to the wakeword recordings directory (ie audio/):\n"))
+
+# wakeword_model_name = str(input("Please enter the name you want to give the wakeword model (ie. 'wakeword_model'):\n"))
+
+# TODO: write data requirements for each flow (include urls)
+
+
+def interactive():
+    cli_text = "Please enter your choice\n1. Optimally split and create a base model from wake-word-recorder data\n2. Generate extra data\n3. Do it all\n\n"
+
+    with open("data_prep_user_configuration.json", "r") as file:
+        user_configuration_dictionary = json.load(file)
+    source_directory = user_configuration_dictionary["audio_source_directory"]
+    wakeword_model_name = user_configuration_dictionary["wakeword_model_name"]
+    pdsounds_directory = user_configuration_dictionary["pdsounds_directory"]
+    directories_to_process = user_configuration_dictionary[
+        "extra_audio_directories_to_process"
+    ]
+    extra_audio_directories_labels = user_configuration_dictionary[
+        "extra_audio_directories_labels"
+    ]
+    max_files_from_source_directory = user_configuration_dictionary[
+        "max_files_from_source_directory"
+    ]
+    max_files_per_destination_directory = user_configuration_dictionary[
+        "max_files_per_destination_directory"
+    ]
+
+    with open("data_prep_system_configuration.json", "r") as file:
+        system_configuration_dictionary = json.load(file)
+    random_split_directories = system_configuration_dictionary[
+        "random_split_directories"
+    ]
+    even_odd_split_directories = system_configuration_dictionary[
+        "even_odd_split_directories"
+    ]
+    three_four_split_directories = system_configuration_dictionary[
+        "three_four_split_directories"
+    ]
+    root_model_name = system_configuration_dictionary["root_model_name"]
+    source_directories = system_configuration_dictionary["source_directories"]
+    destination_directories = system_configuration_dictionary["destination_directories"]
+    directories_to_gauss = system_configuration_dictionary["directories_to_gauss"]
+
+    dialog_handler_instance = DialogHandler("dialog.json", "main_menu_dialog")
+
+    while True:
+        cli_choice = input(
+            dialog_handler_instance.render_template("input-numbered-main_choice")
+        )
+
+        if cli_choice == "1":
+            # TODO: test this flow
+            base_model_info = get_base_model_flow(
+                source_directory,
+                random_split_directories,
+                even_odd_split_directories,
+                three_four_split_directories,
+                root_model_name,
+                wakeword_model_name,
+            )
+
+        elif cli_choice == "2":
+            # TODO: fix base_model_info, it doesn't show up in the flow
+            if base_model_info is None:
+                data_generation_flow(
+                    wakeword_model_name,
+                    source_directories,
+                    destination_directories,
+                    directories_to_gauss,
+                    pdsounds_directory,
+                    base_model_info=None,
+                )
+            else:
+                data_generation_flow(
+                    wakeword_model_name,
+                    source_directories,
+                    destination_directories,
+                    directories_to_gauss,
+                    pdsounds_directory,
+                    base_model_info,
+                )
+
+        elif cli_choice == "3":
+            further_data_generation_flow(
+                directories_to_process,
+                extra_audio_directories_labels,
+                max_files_from_source_directory,
+                max_files_per_destination_directory,
+                wakeword_model_name,
+            )
+
+        elif cli_choice == "4":
+            get_base_model_flow(
+                source_directory,
+                random_split_directories,
+                even_odd_split_directories,
+                three_four_split_directories,
+                root_model_name,
+                wakeword_model_name,
+            )
+            data_generation_flow(
+                wakeword_model_name,
+                source_directories,
+                destination_directories,
+                directories_to_gauss,
+                pdsounds_directory,
+            )
+            further_data_generation_flow(
+                directories_to_process,
+                extra_audio_directories_labels,
+                max_files_from_source_directory,
+                max_files_per_destination_directory,
+                wakeword_model_name,
+            )
+            sys.exit()
+
+        elif cli_choice == "5":
+            sys.exit()
+
+
+"""
+Flow B:
+Requires:
+* Flow A
+* Background noise sound files: 
+    * link: http://pdsounds.tuxfamily.org/
+    * source directory path
+    * convert dataset?
+
+Performs:
+* Data generation: generates both gaussian noise and mixes background noises into samples
+"""
+
+"""
+Flow C:
+Requires:
+* Flow A
+* Incremental 'not-wake-word' test-training datasets: (links) source directory path
+    * TODO: sample and split dataset (common voice dataset)
+    * TODO: import: https://archive.org/download/OpenPathMusic44V5/OpenPathMusic44V5.zip
+    * TODO: convert dataset (mp3 to wav with correct sample rate)
+
+Performs:
+* Incremental test-traing: 
+    * tests the current model against random recordings
+    * if a recording triggers the classifier as 'wake-word' it is chopped into pieces and added into the test and training test (80/20 random split)
+    * the current model is deleted and a new model with the new generated training and test data is trained
+    * the process repeats until all of the incremental 'not-wake-word' test-training datasets have been run.
+
+Output:
+Final production model to be used with the Precise engine. 
+Note: Currently just the tf1 model is used. In the next version, the tf1 model will be converted to tf2 (tflite).
+
+"""
