+++ src/data_prep/data_generation_flow.py
@@ -1,4 +1,4 @@
-from wake_word_data_prep_classes import (
+from data_prep.precise_ops import (
     BasicFileOperations,
     PreciseModelingOperations,
     GaussianNoiseHandler,
