+++ src/data_prep/wake_word_data_prep_flows.py
@@ -1,4 +1,9 @@
-from wake_word_data_prep_classes import BasicFileOperations, TrainTestSplit, PreciseModelingOperations, GaussianNoiseHandler
+from data_prep.precise_ops import (
+    BasicFileOperations,
+    TrainTestSplit,
+    PreciseModelingOperations,
+    GaussianNoiseHandler,
+)
 
 
 train_test_split_instance = TrainTestSplit()
@@ -6,12 +11,12 @@ precise_modeling_operations_instance = PreciseModelingOperations()
 basic_file_operations = BasicFileOperations()
 gaussian_handler_instance = GaussianNoiseHandler()
 
-#TODO: The flows will be used in a CLI, where the varibles can be given in the command line
+# TODO: The flows will be used in a CLI, where the varibles can be given in the command line
 
 
-#TODO: Where do I put in the functions for cleaning up after 
+# TODO: Where do I put in the functions for cleaning up after
 
-#TODO: Where does incremental training (and precise listen, precise test) come in? Is it in another class?
+# TODO: Where does incremental training (and precise listen, precise test) come in? Is it in another class?
 # this final flow!
 # A model is trained from the 'base' data
 # precise test, listen (5 times in a row wake word)
@@ -29,7 +34,7 @@ gaussian_handler_instance = GaussianNoiseHandler()
 # ... And hopefully after common-voice is done, the last retrained model will be dope during the test and listen!
 
 
-'''
+"""
 #Flow A: 
 # User inputs
 # Note: user gives source directory
@@ -78,15 +83,15 @@ precise_modeling_operations_instance.delete_experiment_directories(selected_mode
 
 # When done with incremental, delete all the models, keep name/directory of selected_model_name, remix generated data, re-train
 #precise_modeling_operations_instance.delete_experiment_models()
-'''
+"""
 
 # Delete model files:
-'''
+"""
 precise_modeling_operations_instance = PreciseModelingOperations()
 precise_modeling_operations_instance.delete_model('experiment_1')
-'''
+"""
 
-'''
+"""
 # Flow B: incremental over random
 # Before gaussian noise, run an incremental over the random recordings (TV and conversation)
 # Sub flow: combine the test and training from the random directory results
@@ -136,9 +141,9 @@ destination_directories = [
         'test/wake-word/background_noise/variations/',
     ]
 precise_modeling_operations_instance.move_noise_directories(model_name, source_directories, destination_directories)
-'''
+"""
 # Flow D: Gauss
-'''
+"""
 model_name = 'auto_generated_jarvis'
 
 directories_to_gauss = [
@@ -153,14 +158,14 @@ directories_to_gauss = [
 ]
 
 gaussian_handler_instance.add_gaussian_noise_to_directories(directories_to_gauss)
-'''
+"""
 
 # Flow E: The final flow
 # train a fresh model from the data (450 epochs? -> add epoch parameter in training method)
 # Should the number of epochs grow with the number of files?
 # This code will need to be more flexible to collect the directories from any user
-model_name = 'auto_generated_jarvis'
-'''epochs = '450'
+model_name = "auto_generated_jarvis"
+"""epochs = '450'
 print('Training new model from collected data')
 precise_modeling_operations_instance.run_precise_train(model_name, epochs)
 
@@ -176,16 +181,16 @@ incremental_data_directories = [
 
 # incremental on directories, re-split generated, delete old model, and retrain model 
 precise_modeling_operations_instance.multi_incremental_training(model_name, incremental_data_directories)
-'''
-'''
+"""
+"""
 print('Training final model')
 epochs = '800'
 precise_modeling_operations_instance.run_precise_train(model_name, epochs)
-'''
+"""
 # precise-listen, precise test ?
 # TODO: add in feature for precise-test and parse the output (remove any gauss or added background noise from %)
 # repeat until all directories done
-# TODO: add in flows for sound data wrangling 
+# TODO: add in flows for sound data wrangling
 # getting a max amount of files from a huge directory, break them down into directories with max files per directory
 # converting mp3s to wav files in a directory
 
@@ -194,28 +199,28 @@ precise_modeling_operations_instance.run_precise_train(model_name, epochs)
 
 # Flow 1: get files and copy them to a directory
 # turn into function?
-'''
+"""
 source_directory = "jarvis_S_B_remixed/jarvis_S_B_data_prepped/not-wake-word/parts/"
 destination_directory = 'jarvis_S_B_remixed/jarvis_S_B_data_prepped/not-wake-word/dummy/'
 
 basic_file_operations_instance = BasicFileOperations(source_directory)
 files = basic_file_operations_instance.get_files()
 basic_file_operations_instance.copy_directory(files, destination_directory)
-'''
+"""
 
 # Flow 2: combine files from several (2 or more) directories and copy them to a directory
 # function?
-'''
+"""
 source_directories = ["jarvis_S_B_remixed/jarvis_S_B_data_prepped/not-wake-word/generated/", "jarvis_S_B_remixed/jarvis_S_B_data_prepped/test/not-wake-word/generated/"]
 destination_directory = 'jarvis_S_B_remixed/generated/'
 for source_directory in source_directories:
     basic_file_operations_instance = BasicFileOperations(source_directory)
     files = basic_file_operations_instance.get_files()
     basic_file_operations_instance.copy_directory(files, destination_directory)
-'''
+"""
 
 # Flow 3: train test split random
-'''
+"""
 source_directory = "jarvis_S_B_remixed/generated/"
 training_directory = "jarvis_S_B_remixed/generated/training/"
 testing_directory = "jarvis_S_B_remixed/generated/testing/"
@@ -227,10 +232,10 @@ train_test_split_instance = TrainTestSplit()
 random_selected_training_files, random_selected_testing_files = train_test_split_instance.random_training_test_split(files, dataset_percent_size)
 basic_file_operations_instance.copy_directory(random_selected_training_files, training_directory)
 basic_file_operations_instance.copy_directory(random_selected_testing_files, testing_directory)
-'''
+"""
 
 # Flow 4: train test split even-odd
-'''
+"""
 source_directory = "jarvis_S_B_remixed/not-wake-word/parts/"
 training_directory = "jarvis_S_B_remixed/not-wake-word/parts/training/"
 testing_directory = "jarvis_S_B_remixed/not-wake-word/parts/testing/"
@@ -242,10 +247,10 @@ train_test_split_instance = TrainTestSplit()
 selected_training_files, selected_testing_files = train_test_split_instance.even_odd_training_test_split(files)
 basic_file_operations_instance.copy_directory(selected_training_files, training_directory)
 basic_file_operations_instance.copy_directory(selected_testing_files, testing_directory)
-'''
+"""
 
 # Flow 5: split_directory (by type)
-'''
+"""
 source_directory = 'jarvis_S_B_remixed/not-wake-word/parts/'
 training_directory = source_directory+'training/'
 testing_directory = source_directory+'testing/'
@@ -253,10 +258,10 @@ split_type = 'even-odd'
 
 train_test_split_instance = TrainTestSplit()
 train_test_split_instance.split_directory(source_directory, training_directory, testing_directory, split_type)
-'''
+"""
 
 # Flow 6: split_multiple_directories:
-'''
+"""
 model_name = 'experiment'
 source_directory = 'jarvis_S_B_remixed/'
 
@@ -279,21 +284,26 @@ even_odd_split_directories = [
 
 train_test_split_instance = TrainTestSplit()
 train_test_split_instance.split_multiple_directories(model_name)
-'''
+"""
 # Flow 7: precise train
-'''
+"""
 # sub directories don't work so well here yet! Just use 1 directory
 model_name = 'experiment_1'
 training_run = precise_modeling_operations_instance.run_precise_train(model_name)
 precise_modeling_operations_instance.get_last_epoch_model_info(model_name, training_run)
 print(precise_modeling_operations_instance.models)
-'''
+"""
 
 # TEST STUFF
 
-source_directory = 'jarvis_S_B_remixed/'
+source_directory = "jarvis_S_B_remixed/"
 
 # Note: user either selects default directories or can add their own directories
-random_split_directory = 'not-wake-word/test/'
+random_split_directory = "not-wake-word/test/"
 
-train_test_split_instance.split_directory(source_directory + random_split_directory, source_directory + 'test_train/', source_directory + 'test_test/', 'three_four')
+train_test_split_instance.split_directory(
+    source_directory + random_split_directory,
+    source_directory + "test_train/",
+    source_directory + "test_test/",
+    "three_four",
+)
